% Generated by build script
\name{nws_forecast}
\alias{nws_forecast}
\title{Get weather forecast for a location}
\usage{
nws_forecast()
}
\arguments{
\item{latitude}{Latitude (e.g. 40.7128 for NYC)}
\item{longitude}{Longitude (e.g. -74.0060 for NYC)}
}
\value{tibble: number, name (period name), temperature, temperature_unit,}
\description{Get weather forecast for a location}
