% Generated by build script
\name{nws_stations}
\alias{nws_stations}
\title{Get observation stations near a location}
\usage{
nws_stations()
}
\arguments{
\item{latitude}{Latitude}
\item{longitude}{Longitude}
}
\value{tibble: station_id, name, latitude, longitude, elevation_m}
\description{Get observation stations near a location}
